move drop to a file
