+++ tests/test_cnn/test_transformer.py
@@ -1,7 +1,8 @@
 import pytest
 import torch
 
-from mmcv.cnn.bricks.transformer import (FFN, BaseTransformerLayer, DropPath,
+from mmcv.cnn.bricks.drop import DropPath
+from mmcv.cnn.bricks.transformer import (FFN, BaseTransformerLayer,
                                          MultiheadAttention,
                                          TransformerLayerSequence)
